+++ payments-core/src/main/java/com/stripe/android/payments/core/authentication/threeds2/Stripe3DS2NextActionHandler.kt
@@ -61,7 +61,8 @@ internal class Stripe3DS2NextActionHandler @Inject constructor(
     override suspend fun performNextActionOnResumed(
         host: AuthActivityStarterHost,
         actionable: StripeIntent,
-        requestOptions: ApiRequest.Options
+        requestOptions: ApiRequest.Options,
+        returnUrl: String?
     ) {
         stripe3ds2CompletionStarterFactory(host).start(
             Stripe3ds2TransactionContract.Args(